pyproject.toml
[tool.poetry]
name = "kakawa-rag-api"
version = "0.1.1"
package-mode = false
description = ""
authors = ["Luandro <luandro@digital-democracy.org>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
pydantic = "^1.8.2"
llama-index = "^0.10.18"
datasets = "^2.18.0"
ragas = "^0.1.3"
pandas = "^2.2.1"
langfuse = "^2.20.0"
python-dotenv = "^1.0.1"
fastapi = "^0.110.0"
uvicorn = "^0.28.0"
llama-index-llms-ollama = "^0.1.2"
llama-index-core = "^0.10.18.post1"
qdrant-client = "^1.8.0"
llama-index-vector-stores-qdrant = "^0.1.4"
llama-index-embeddings-openai = "^0.1.6"
llama-index-embeddings-cohere = "^0.1.3"
llama-index-embeddings-fastembed = "^0.1.4"
fastapi-offline = "^1.7.1"
llama-index-postprocessor-colbert-rerank = "^0.1.1"
spacy = "^3.7.4"
span-marker = "^1.5.0"
scikit-learn = "^1.4.1.post1"
llama-index-llms-together = "^0.1.3"
llama-index-llms-groq = "^0.1.3"
beautifulsoup4 = "^4.12.3"
llama-index-llms-anthropic = "^0.1.7"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
pre-commit = "^3.6.2"
black = "^24.2.0"
isort = "^5.13.2"

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
